+++ src/routes/routes.ts
@@ -762,6 +762,7 @@ function routes(logger, app: express.Application, spinalAPIMiddleware: ISpinalAP
   require('./groupContext/group/updateGroup')(logger, app, spinalAPIMiddleware);
   require('./groupContext/group/deleteGroup')(logger, app, spinalAPIMiddleware);
   require('./groupContext/group/readGroup')(logger, app, spinalAPIMiddleware);
+  require('./groupContext/group/unassignItem')(logger, app, spinalAPIMiddleware);
 
   // notes routes
   require('./notes/addNotes')(logger, app, spinalAPIMiddleware);